nifi_rs::api

Struct ParameterProviderDto

pub struct ParameterProviderDto {

    pub affected_components: Option<Vec<AffectedComponentEntity>>,
    pub annotation_data: Option<String>,
    pub bundle: Option<BundleDto>,
    pub comments: Option<String>,
    pub custom_ui_url: Option<String>,
    pub deprecated: Option<bool>,
    pub descriptors: HashMap<String, PropertyDescriptorDto>,
    pub extension_missing: Option<bool>,
    pub id: Option<String>,
    pub multiple_versions_available: Option<bool>,
    pub name: Option<String>,
    pub parameter_group_configurations: Vec<ParameterGroupConfigurationEntity>,
    pub parameter_status: Option<Vec<ParameterStatusDto>>,
    pub parent_group_id: Option<String>,
    pub persists_state: Option<bool>,
    pub position: Option<PositionDto>,
    pub properties: HashMap<String, String>,
    pub referencing_parameter_contexts: Option<Vec<ParameterProviderReferencingComponentEntity>>,
    pub restricted: Option<bool>,
    pub type_: Option<String>,
    pub validation_errors: Vec<String>,
    pub validation_status: Option<ParameterProviderDtoValidationStatus>,
    pub versioned_component_id: Option<String>,

}

ParameterProviderDto

JSON schema

{
 "type": "object",
 "properties": {
   "affectedComponents": {
     "description": "The set of all components in the flow that are referencing Parameters provided by this provider",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/AffectedComponentEntity"
     },
     "uniqueItems": true
   },
   "annotationData": {
     "description": "The annotation data for the parameter provider. This is how the custom UI relays configuration to the parameter provider.",
     "type": "string"
   },
   "bundle": {
     "$ref": "#/components/schemas/BundleDTO"
   },
   "comments": {
     "description": "The comments of the parameter provider.",
     "type": "string"
   },
   "customUiUrl": {
     "description": "The URL for the custom configuration UI for the parameter provider.",
     "type": "string"
   },
   "deprecated": {
     "description": "Whether the parameter provider has been deprecated.",
     "type": "boolean"
   },
   "descriptors": {
     "description": "The descriptors for the parameter providers properties.",
     "type": "object",
     "additionalProperties": {
       "$ref": "#/components/schemas/PropertyDescriptorDTO"
     }
   },
   "extensionMissing": {
     "description": "Whether the underlying extension is missing.",
     "type": "boolean"
   },
   "id": {
     "description": "The id of the component.",
     "type": "string"
   },
   "multipleVersionsAvailable": {
     "description": "Whether the parameter provider has multiple versions available.",
     "type": "boolean"
   },
   "name": {
     "description": "The name of the parameter provider.",
     "type": "string"
   },
   "parameterGroupConfigurations": {
     "description": "Configuration for any fetched parameter groups.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/ParameterGroupConfigurationEntity"
     }
   },
   "parameterStatus": {
     "description": "The status of all provided parameters for this parameter provider",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/ParameterStatusDTO"
     },
     "uniqueItems": true
   },
   "parentGroupId": {
     "description": "The id of parent process group of this component if applicable.",
     "type": "string"
   },
   "persistsState": {
     "description": "Whether the parameter provider persists state.",
     "type": "boolean"
   },
   "position": {
     "$ref": "#/components/schemas/PositionDTO"
   },
   "properties": {
     "description": "The properties of the parameter provider.",
     "type": "object",
     "additionalProperties": {
       "description": "The properties of the parameter provider.",
       "type": "string"
     }
   },
   "referencingParameterContexts": {
     "description": "The Parameter Contexts that reference this Parameter Provider",
     "readOnly": true,
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/ParameterProviderReferencingComponentEntity"
     },
     "uniqueItems": true
   },
   "restricted": {
     "description": "Whether the parameter provider requires elevated privileges.",
     "type": "boolean"
   },
   "type": {
     "description": "The fully qualified type of the parameter provider.",
     "type": "string"
   },
   "validationErrors": {
     "description": "Gets the validation errors from the parameter provider. These validation errors represent the problems with the parameter provider that must be resolved before it can be scheduled to run.",
     "type": "array",
     "items": {
       "description": "Gets the validation errors from the parameter provider. These validation errors represent the problems with the parameter provider that must be resolved before it can be scheduled to run.",
       "type": "string"
     }
   },
   "validationStatus": {
     "description": "Indicates whether the Parameter Provider is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Parameter Provider is valid)",
     "readOnly": true,
     "type": "string",
     "enum": [
       "VALID",
       "INVALID",
       "VALIDATING"
     ]
   },
   "versionedComponentId": {
     "description": "The ID of the corresponding component that is under version control",
     "type": "string"
   }
 }
}

Fields
affected_components: Option<Vec<AffectedComponentEntity>>

The set of all components in the flow that are referencing Parameters provided by this provider

annotation_data: Option<String>

The annotation data for the parameter provider. This is how the custom UI relays configuration to the parameter provider.

bundle: Option<BundleDto>comments: Option<String>

The comments of the parameter provider.

custom_ui_url: Option<String>

The URL for the custom configuration UI for the parameter provider.

deprecated: Option<bool>

Whether the parameter provider has been deprecated.

descriptors: HashMap<String, PropertyDescriptorDto>

The descriptors for the parameter providers properties.

extension_missing: Option<bool>

Whether the underlying extension is missing.

id: Option<String>

The id of the component.

multiple_versions_available: Option<bool>

Whether the parameter provider has multiple versions available.

name: Option<String>

The name of the parameter provider.

parameter_group_configurations: Vec<ParameterGroupConfigurationEntity>

Configuration for any fetched parameter groups.

parameter_status: Option<Vec<ParameterStatusDto>>

The status of all provided parameters for this parameter provider

parent_group_id: Option<String>

The id of parent process group of this component if applicable.

persists_state: Option<bool>

Whether the parameter provider persists state.

position: Option<PositionDto>properties: HashMap<String, String>

The properties of the parameter provider.

referencing_parameter_contexts: Option<Vec<ParameterProviderReferencingComponentEntity>>

The Parameter Contexts that reference this Parameter Provider

restricted: Option<bool>

Whether the parameter provider requires elevated privileges.

type_: Option<String>

The fully qualified type of the parameter provider.

validation_errors: Vec<String>

Gets the validation errors from the parameter provider. These validation errors represent the problems with the parameter provider that must be resolved before it can be scheduled to run.

validation_status: Option<ParameterProviderDtoValidationStatus>

Indicates whether the Parameter Provider is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Parameter Provider is valid)

versioned_component_id: Option<String>

The ID of the corresponding component that is under version control

Implementations

impl ParameterProviderDto

pub fn builder() -> ParameterProviderDto