nifi_rs::api

Struct ProcessorConfigDto

pub struct ProcessorConfigDto {

    pub annotation_data: Option<String>,
    pub auto_terminated_relationships: Option<Vec<String>>,
    pub backoff_mechanism: Option<String>,
    pub bulletin_level: Option<String>,
    pub comments: Option<String>,
    pub concurrently_schedulable_task_count: Option<i32>,
    pub custom_ui_url: Option<String>,
    pub default_concurrent_tasks: HashMap<String, String>,
    pub default_scheduling_period: HashMap<String, String>,
    pub descriptors: HashMap<String, PropertyDescriptorDto>,
    pub execution_node: Option<String>,
    pub loss_tolerant: Option<bool>,
    pub max_backoff_period: Option<String>,
    pub penalty_duration: Option<String>,
    pub properties: HashMap<String, String>,
    pub retried_relationships: Option<Vec<String>>,
    pub retry_count: Option<i32>,
    pub run_duration_millis: Option<i64>,
    pub scheduling_period: Option<String>,
    pub scheduling_strategy: Option<String>,
    pub sensitive_dynamic_property_names: Option<Vec<String>>,
    pub yield_duration: Option<String>,

}

The configuration details for the processor. These details will be included in a response if the verbose flag is included in a request.

JSON schema

{
 "description": "The configuration details for the processor. These details will be included in a response if the verbose flag is included in a request.",
 "type": "object",
 "properties": {
   "annotationData": {
     "description": "The annotation data for the processor used to relay configuration between a custom UI and the procesosr.",
     "type": "string"
   },
   "autoTerminatedRelationships": {
     "description": "The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the 'isAutoTerminate' property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.",
     "type": "array",
     "items": {
       "description": "The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the 'isAutoTerminate' property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.",
       "type": "string"
     },
     "uniqueItems": true
   },
   "backoffMechanism": {
     "description": "Determines whether the FlowFile should be penalized or the processor should be yielded between retries. Possible returned values: PENALIZE_FLOWFILE, YIELD_PROCESSOR. See BackoffMechanism.class for more details.",
     "type": "string"
   },
   "bulletinLevel": {
     "description": "The level at which the processor will report bulletins.",
     "type": "string"
   },
   "comments": {
     "description": "The comments for the processor.",
     "type": "string"
   },
   "concurrentlySchedulableTaskCount": {
     "description": "The number of tasks that should be concurrently schedule for the processor. If the processor doesn't allow parallol processing then any positive input will be ignored.",
     "type": "integer",
     "format": "int32"
   },
   "customUiUrl": {
     "description": "The URL for the processor's custom configuration UI if applicable.",
     "type": "string"
   },
   "defaultConcurrentTasks": {
     "description": "Maps default values for concurrent tasks for each applicable scheduling strategy.",
     "type": "object",
     "additionalProperties": {
       "description": "Maps default values for concurrent tasks for each applicable scheduling strategy.",
       "type": "string"
     }
   },
   "defaultSchedulingPeriod": {
     "description": "Maps default values for scheduling period for each applicable scheduling strategy.",
     "type": "object",
     "additionalProperties": {
       "description": "Maps default values for scheduling period for each applicable scheduling strategy.",
       "type": "string"
     }
   },
   "descriptors": {
     "description": "Descriptors for the processor's properties.",
     "type": "object",
     "additionalProperties": {
       "$ref": "#/components/schemas/PropertyDescriptorDTO"
     }
   },
   "executionNode": {
     "description": "Indicates the node where the process will execute.",
     "type": "string"
   },
   "lossTolerant": {
     "description": "Whether the processor is loss tolerant.",
     "type": "boolean"
   },
   "maxBackoffPeriod": {
     "description": "Maximum amount of time to be waited during a retry period.",
     "type": "string"
   },
   "penaltyDuration": {
     "description": "The amount of time that is used when the process penalizes a flowfile.",
     "type": "string"
   },
   "properties": {
     "description": "The properties for the processor. Properties whose value is not set will only contain the property name.",
     "type": "object",
     "additionalProperties": {
       "description": "The properties for the processor. Properties whose value is not set will only contain the property name.",
       "type": "string"
     }
   },
   "retriedRelationships": {
     "description": "All the relationships should be retried.",
     "type": "array",
     "items": {
       "description": "All the relationships should be retried.",
       "type": "string"
     },
     "uniqueItems": true
   },
   "retryCount": {
     "description": "Overall number of retries.",
     "type": "integer",
     "format": "int32"
   },
   "runDurationMillis": {
     "description": "The run duration for the processor in milliseconds.",
     "type": "integer",
     "format": "int64"
   },
   "schedulingPeriod": {
     "description": "The frequency with which to schedule the processor. The format of the value will depend on th value of schedulingStrategy.",
     "type": "string"
   },
   "schedulingStrategy": {
     "description": "Indicates how the processor should be scheduled to run.",
     "type": "string"
   },
   "sensitiveDynamicPropertyNames": {
     "description": "Set of sensitive dynamic property names",
     "type": "array",
     "items": {
       "description": "Set of sensitive dynamic property names",
       "type": "string"
     },
     "uniqueItems": true
   },
   "yieldDuration": {
     "description": "The amount of time that must elapse before this processor is scheduled again after yielding.",
     "type": "string"
   }
 }
}

Fields
annotation_data: Option<String>

The annotation data for the processor used to relay configuration between a custom UI and the procesosr.

auto_terminated_relationships: Option<Vec<String>>

The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the ‘isAutoTerminate’ property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.

backoff_mechanism: Option<String>

Determines whether the FlowFile should be penalized or the processor should be yielded between retries. Possible returned values: PENALIZE_FLOWFILE, YIELD_PROCESSOR. See BackoffMechanism.class for more details.

bulletin_level: Option<String>

The level at which the processor will report bulletins.

comments: Option<String>

The comments for the processor.

concurrently_schedulable_task_count: Option<i32>

The number of tasks that should be concurrently schedule for the processor. If the processor doesn’t allow parallol processing then any positive input will be ignored.

custom_ui_url: Option<String>

The URL for the processor’s custom configuration UI if applicable.

default_concurrent_tasks: HashMap<String, String>

Maps default values for concurrent tasks for each applicable scheduling strategy.

default_scheduling_period: HashMap<String, String>

Maps default values for scheduling period for each applicable scheduling strategy.

descriptors: HashMap<String, PropertyDescriptorDto>

Descriptors for the processor’s properties.

execution_node: Option<String>

Indicates the node where the process will execute.

loss_tolerant: Option<bool>

Whether the processor is loss tolerant.

max_backoff_period: Option<String>

Maximum amount of time to be waited during a retry period.

penalty_duration: Option<String>

The amount of time that is used when the process penalizes a flowfile.

properties: HashMap<String, String>

The properties for the processor. Properties whose value is not set will only contain the property name.

retried_relationships: Option<Vec<String>>

All the relationships should be retried.

retry_count: Option<i32>

Overall number of retries.

run_duration_millis: Option<i64>

The run duration for the processor in milliseconds.

scheduling_period: Option<String>

The frequency with which to schedule the processor. The format of the value will depend on th value of schedulingStrategy.

scheduling_strategy: Option<String>

Indicates how the processor should be scheduled to run.

sensitive_dynamic_property_names: Option<Vec<String>>

Set of sensitive dynamic property names

yield_duration: Option<String>

The amount of time that must elapse before this processor is scheduled again after yielding.

Implementations

impl ProcessorConfigDto

pub fn builder() -> ProcessorConfigDto