nifi_rs::api

Struct SystemDiagnosticsSnapshotDto

pub struct SystemDiagnosticsSnapshotDto {

    pub available_processors: Option<i32>,
    pub content_repository_storage_usage: Option<Vec<StorageUsageDto>>,
    pub daemon_threads: Option<i32>,
    pub flow_file_repository_storage_usage: Option<StorageUsageDto>,
    pub free_heap: Option<String>,
    pub free_heap_bytes: Option<i64>,
    pub free_non_heap: Option<String>,
    pub free_non_heap_bytes: Option<i64>,
    pub garbage_collection: Option<Vec<GarbageCollectionDto>>,
    pub heap_utilization: Option<String>,
    pub max_heap: Option<String>,
    pub max_heap_bytes: Option<i64>,
    pub max_non_heap: Option<String>,
    pub max_non_heap_bytes: Option<i64>,
    pub non_heap_utilization: Option<String>,
    pub processor_load_average: Option<f64>,
    pub provenance_repository_storage_usage: Option<Vec<StorageUsageDto>>,
    pub resource_claim_details: Vec<ResourceClaimDetailsDto>,
    pub stats_last_refreshed: Option<String>,
    pub total_heap: Option<String>,
    pub total_heap_bytes: Option<i64>,
    pub total_non_heap: Option<String>,
    pub total_non_heap_bytes: Option<i64>,
    pub total_threads: Option<i32>,
    pub uptime: Option<String>,
    pub used_heap: Option<String>,
    pub used_heap_bytes: Option<i64>,
    pub used_non_heap: Option<String>,
    pub used_non_heap_bytes: Option<i64>,
    pub version_info: Option<VersionInfoDto>,

}

The System Diagnostics snapshot from the node.

JSON schema

{
 "description": "The System Diagnostics snapshot from the node.",
 "type": "object",
 "properties": {
   "availableProcessors": {
     "description": "Number of available processors if supported by the underlying system.",
     "type": "integer",
     "format": "int32"
   },
   "contentRepositoryStorageUsage": {
     "description": "The content repository storage usage.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/StorageUsageDTO"
     },
     "uniqueItems": true
   },
   "daemonThreads": {
     "description": "Number of daemon threads.",
     "type": "integer",
     "format": "int32"
   },
   "flowFileRepositoryStorageUsage": {
     "$ref": "#/components/schemas/StorageUsageDTO"
   },
   "freeHeap": {
     "description": "Amount of free heap.",
     "type": "string"
   },
   "freeHeapBytes": {
     "description": "The number of bytes that are allocated to the JVM heap but not currently being used",
     "type": "integer",
     "format": "int64"
   },
   "freeNonHeap": {
     "description": "Amount of free non heap.",
     "type": "string"
   },
   "freeNonHeapBytes": {
     "description": "Total number of free non-heap bytes available to the JVM",
     "type": "integer",
     "format": "int64"
   },
   "garbageCollection": {
     "description": "The garbage collection details.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/GarbageCollectionDTO"
     },
     "uniqueItems": true
   },
   "heapUtilization": {
     "description": "Utilization of heap.",
     "type": "string"
   },
   "maxHeap": {
     "description": "Maximum size of heap.",
     "type": "string"
   },
   "maxHeapBytes": {
     "description": "The maximum number of bytes that can be used by the JVM",
     "type": "integer",
     "format": "int64"
   },
   "maxNonHeap": {
     "description": "Maximum size of non heap.",
     "type": "string"
   },
   "maxNonHeapBytes": {
     "description": "The maximum number of bytes that the JVM can use for non-heap purposes",
     "type": "integer",
     "format": "int64"
   },
   "nonHeapUtilization": {
     "description": "Utilization of non heap.",
     "type": "string"
   },
   "processorLoadAverage": {
     "description": "The processor load average if supported by the underlying system.",
     "type": "number",
     "format": "double"
   },
   "provenanceRepositoryStorageUsage": {
     "description": "The provenance repository storage usage.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/StorageUsageDTO"
     },
     "uniqueItems": true
   },
   "resourceClaimDetails": {
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/ResourceClaimDetailsDTO"
     }
   },
   "statsLastRefreshed": {
     "description": "When the diagnostics were generated.",
     "type": "string"
   },
   "totalHeap": {
     "description": "Total size of heap.",
     "type": "string"
   },
   "totalHeapBytes": {
     "description": "The total number of bytes that are available for the JVM heap to use",
     "type": "integer",
     "format": "int64"
   },
   "totalNonHeap": {
     "description": "Total size of non heap.",
     "type": "string"
   },
   "totalNonHeapBytes": {
     "description": "Total number of bytes allocated to the JVM not used for heap",
     "type": "integer",
     "format": "int64"
   },
   "totalThreads": {
     "description": "Total number of threads.",
     "type": "integer",
     "format": "int32"
   },
   "uptime": {
     "description": "The uptime of the Java virtual machine",
     "type": "string"
   },
   "usedHeap": {
     "description": "Amount of used heap.",
     "type": "string"
   },
   "usedHeapBytes": {
     "description": "The number of bytes of JVM heap that are currently being used",
     "type": "integer",
     "format": "int64"
   },
   "usedNonHeap": {
     "description": "Amount of use non heap.",
     "type": "string"
   },
   "usedNonHeapBytes": {
     "description": "Total number of bytes used by the JVM not in the heap space",
     "type": "integer",
     "format": "int64"
   },
   "versionInfo": {
     "$ref": "#/components/schemas/VersionInfoDTO"
   }
 }
}

Fields
available_processors: Option<i32>

Number of available processors if supported by the underlying system.

content_repository_storage_usage: Option<Vec<StorageUsageDto>>

The content repository storage usage.

daemon_threads: Option<i32>

Number of daemon threads.

flow_file_repository_storage_usage: Option<StorageUsageDto>free_heap: Option<String>

Amount of free heap.

free_heap_bytes: Option<i64>

The number of bytes that are allocated to the JVM heap but not currently being used

free_non_heap: Option<String>

Amount of free non heap.

free_non_heap_bytes: Option<i64>

Total number of free non-heap bytes available to the JVM

garbage_collection: Option<Vec<GarbageCollectionDto>>

The garbage collection details.

heap_utilization: Option<String>

Utilization of heap.

max_heap: Option<String>

Maximum size of heap.

max_heap_bytes: Option<i64>

The maximum number of bytes that can be used by the JVM

max_non_heap: Option<String>

Maximum size of non heap.

max_non_heap_bytes: Option<i64>

The maximum number of bytes that the JVM can use for non-heap purposes

non_heap_utilization: Option<String>

Utilization of non heap.

processor_load_average: Option<f64>provenance_repository_storage_usage: Option<Vec<StorageUsageDto>>

The provenance repository storage usage.

resource_claim_details: Vec<ResourceClaimDetailsDto>stats_last_refreshed: Option<String>

When the diagnostics were generated.

total_heap: Option<String>

Total size of heap.

total_heap_bytes: Option<i64>

The total number of bytes that are available for the JVM heap to use

total_non_heap: Option<String>

Total size of non heap.

total_non_heap_bytes: Option<i64>

Total number of bytes allocated to the JVM not used for heap

total_threads: Option<i32>

Total number of threads.

uptime: Option<String>

The uptime of the Java virtual machine

used_heap: Option<String>

Amount of used heap.

used_heap_bytes: Option<i64>

The number of bytes of JVM heap that are currently being used

used_non_heap: Option<String>

Amount of use non heap.

used_non_heap_bytes: Option<i64>

Total number of bytes used by the JVM not in the heap space

version_info: Option<VersionInfoDto>
Implementations

impl SystemDiagnosticsSnapshotDto

pub fn builder() -> SystemDiagnosticsSnapshotDto