nifi_rs::api

Struct VersionedLabel

pub struct VersionedLabel {
    pub comments: Option<String>,
    pub component_type: Option<VersionedLabelComponentType>,
    pub group_identifier: Option<String>,
    pub height: Option<f64>,
    pub identifier: Option<String>,
    pub instance_identifier: Option<String>,
    pub label: Option<String>,
    pub name: Option<String>,
    pub position: Option<Position>,
    pub style: HashMap<String, String>,
    pub width: Option<f64>,
    pub z_index: Option<i64>,
}

The Labels

JSON schema

{
 "description": "The Labels",
 "type": "object",
 "properties": {
   "comments": {
     "description": "The user-supplied comments for the component",
     "type": "string"
   },
   "componentType": {
     "type": "string",
     "enum": [
       "CONNECTION",
       "PROCESSOR",
       "PROCESS_GROUP",
       "REMOTE_PROCESS_GROUP",
       "INPUT_PORT",
       "OUTPUT_PORT",
       "REMOTE_INPUT_PORT",
       "REMOTE_OUTPUT_PORT",
       "FUNNEL",
       "LABEL",
       "CONTROLLER_SERVICE",
       "REPORTING_TASK",
       "FLOW_ANALYSIS_RULE",
       "PARAMETER_CONTEXT",
       "PARAMETER_PROVIDER",
       "FLOW_REGISTRY_CLIENT"
     ]
   },
   "groupIdentifier": {
     "description": "The ID of the Process Group that this component belongs to",
     "type": "string"
   },
   "height": {
     "description": "The height of the label in pixels when at a 1:1 scale.",
     "type": "number",
     "format": "double"
   },
   "identifier": {
     "description": "The component's unique identifier",
     "type": "string"
   },
   "instanceIdentifier": {
     "description": "The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component",
     "type": "string"
   },
   "label": {
     "description": "The text that appears in the label.",
     "type": "string"
   },
   "name": {
     "description": "The component's name",
     "type": "string"
   },
   "position": {
     "$ref": "#/components/schemas/Position"
   },
   "style": {
     "description": "The styles for this label (font-size : 12px, background-color : #eee, etc).",
     "type": "object",
     "additionalProperties": {
       "description": "The styles for this label (font-size : 12px, background-color : #eee, etc).",
       "type": "string"
     }
   },
   "width": {
     "description": "The width of the label in pixels when at a 1:1 scale.",
     "type": "number",
     "format": "double"
   },
   "zIndex": {
     "description": "The z index of the connection.",
     "type": "integer",
     "format": "int64"
   }
 }
}

Fields
comments: Option<String>

The user-supplied comments for the component

component_type: Option<VersionedLabelComponentType>group_identifier: Option<String>

The ID of the Process Group that this component belongs to

height: Option<f64>identifier: Option<String>

The componentโs unique identifier

instance_identifier: Option<String>

The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component

label: Option<String>

The text that appears in the label.

name: Option<String>

The componentโs name

position: Option<Position>style: HashMap<String, String>

The styles for this label (font-size : 12px, background-color : #eee, etc).

width: Option<f64>z_index: Option<i64>

The z index of the connection.

Implementations

impl VersionedLabel

pub fn builder() -> VersionedLabel