nifi_rs::api

Struct VersionedProcessor

pub struct VersionedProcessor {

    pub annotation_data: Option<String>,
    pub auto_terminated_relationships: Option<Vec<String>>,
    pub backoff_mechanism: Option<String>,
    pub bulletin_level: Option<String>,
    pub bundle: Option<Bundle>,
    pub comments: Option<String>,
    pub component_type: Option<VersionedProcessorComponentType>,
    pub concurrently_schedulable_task_count: Option<i32>,
    pub execution_node: Option<String>,
    pub group_identifier: Option<String>,
    pub identifier: Option<String>,
    pub instance_identifier: Option<String>,
    pub max_backoff_period: Option<String>,
    pub name: Option<String>,
    pub penalty_duration: Option<String>,
    pub position: Option<Position>,
    pub properties: HashMap<String, String>,
    pub property_descriptors: HashMap<String, VersionedPropertyDescriptor>,
    pub retried_relationships: Option<Vec<String>>,
    pub retry_count: Option<i32>,
    pub run_duration_millis: Option<i64>,
    pub scheduled_state: Option<VersionedProcessorScheduledState>,
    pub scheduling_period: Option<String>,
    pub scheduling_strategy: Option<String>,
    pub style: HashMap<String, String>,
    pub type_: Option<String>,
    pub yield_duration: Option<String>,

}

The Processors

JSON schema

{
 "description": "The Processors",
 "type": "object",
 "properties": {
   "annotationData": {
     "description": "The annotation data for the processor used to relay configuration between a custom UI and the procesosr.",
     "type": "string"
   },
   "autoTerminatedRelationships": {
     "description": "The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the 'isAutoTerminate' property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.",
     "type": "array",
     "items": {
       "description": "The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the 'isAutoTerminate' property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.",
       "type": "string"
     },
     "uniqueItems": true
   },
   "backoffMechanism": {
     "description": "Determines whether the FlowFile should be penalized or the processor should be yielded between retries. Possible returned values: PENALIZE_FLOWFILE, YIELD_PROCESSOR.",
     "type": "string"
   },
   "bulletinLevel": {
     "description": "The level at which the processor will report bulletins.",
     "type": "string"
   },
   "bundle": {
     "$ref": "#/components/schemas/Bundle"
   },
   "comments": {
     "description": "The user-supplied comments for the component",
     "type": "string"
   },
   "componentType": {
     "type": "string",
     "enum": [
       "CONNECTION",
       "PROCESSOR",
       "PROCESS_GROUP",
       "REMOTE_PROCESS_GROUP",
       "INPUT_PORT",
       "OUTPUT_PORT",
       "REMOTE_INPUT_PORT",
       "REMOTE_OUTPUT_PORT",
       "FUNNEL",
       "LABEL",
       "CONTROLLER_SERVICE",
       "REPORTING_TASK",
       "FLOW_ANALYSIS_RULE",
       "PARAMETER_CONTEXT",
       "PARAMETER_PROVIDER",
       "FLOW_REGISTRY_CLIENT"
     ]
   },
   "concurrentlySchedulableTaskCount": {
     "description": "The number of tasks that should be concurrently schedule for the processor. If the processor doesn't allow parallol processing then any positive input will be ignored.",
     "type": "integer",
     "format": "int32"
   },
   "executionNode": {
     "description": "Indicates the node where the process will execute.",
     "type": "string"
   },
   "groupIdentifier": {
     "description": "The ID of the Process Group that this component belongs to",
     "type": "string"
   },
   "identifier": {
     "description": "The component's unique identifier",
     "type": "string"
   },
   "instanceIdentifier": {
     "description": "The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component",
     "type": "string"
   },
   "maxBackoffPeriod": {
     "description": "Maximum amount of time to be waited during a retry period.",
     "type": "string"
   },
   "name": {
     "description": "The component's name",
     "type": "string"
   },
   "penaltyDuration": {
     "description": "The amout of time that is used when the process penalizes a flowfile.",
     "type": "string"
   },
   "position": {
     "$ref": "#/components/schemas/Position"
   },
   "properties": {
     "description": "The properties for the component. Properties whose value is not set will only contain the property name.",
     "type": "object",
     "additionalProperties": {
       "description": "The properties for the component. Properties whose value is not set will only contain the property name.",
       "type": "string"
     }
   },
   "propertyDescriptors": {
     "description": "The property descriptors for the component.",
     "type": "object",
     "additionalProperties": {
       "$ref": "#/components/schemas/VersionedPropertyDescriptor"
     }
   },
   "retriedRelationships": {
     "description": "All the relationships should be retried.",
     "type": "array",
     "items": {
       "description": "All the relationships should be retried.",
       "type": "string"
     },
     "uniqueItems": true
   },
   "retryCount": {
     "description": "Overall number of retries.",
     "type": "integer",
     "format": "int32"
   },
   "runDurationMillis": {
     "description": "The run duration for the processor in milliseconds.",
     "type": "integer",
     "format": "int64"
   },
   "scheduledState": {
     "description": "The scheduled state of the component",
     "type": "string",
     "enum": [
       "ENABLED",
       "DISABLED",
       "RUNNING"
     ]
   },
   "schedulingPeriod": {
     "description": "The frequency with which to schedule the processor. The format of the value will depend on th value of schedulingStrategy.",
     "type": "string"
   },
   "schedulingStrategy": {
     "description": "Indicates how the processor should be scheduled to run.",
     "type": "string"
   },
   "style": {
     "description": "Stylistic data for rendering in a UI",
     "type": "object",
     "additionalProperties": {
       "description": "Stylistic data for rendering in a UI",
       "type": "string"
     }
   },
   "type": {
     "description": "The type of the extension component",
     "type": "string"
   },
   "yieldDuration": {
     "description": "The amount of time that must elapse before this processor is scheduled again after yielding.",
     "type": "string"
   }
 }
}

Fields
annotation_data: Option<String>

The annotation data for the processor used to relay configuration between a custom UI and the procesosr.

auto_terminated_relationships: Option<Vec<String>>

The names of all relationships that cause a flow file to be terminated if the relationship is not connected elsewhere. This property differs from the ‘isAutoTerminate’ property of the RelationshipDTO in that the RelationshipDTO is meant to depict the current configuration, whereas this property can be set in a DTO when updating a Processor in order to change which Relationships should be auto-terminated.

backoff_mechanism: Option<String>

Determines whether the FlowFile should be penalized or the processor should be yielded between retries. Possible returned values: PENALIZE_FLOWFILE, YIELD_PROCESSOR.

bulletin_level: Option<String>

The level at which the processor will report bulletins.

bundle: Option<Bundle>comments: Option<String>

The user-supplied comments for the component

component_type: Option<VersionedProcessorComponentType>concurrently_schedulable_task_count: Option<i32>

The number of tasks that should be concurrently schedule for the processor. If the processor doesn’t allow parallol processing then any positive input will be ignored.

execution_node: Option<String>

Indicates the node where the process will execute.

group_identifier: Option<String>

The ID of the Process Group that this component belongs to

identifier: Option<String>

The component’s unique identifier

instance_identifier: Option<String>

The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component

max_backoff_period: Option<String>

Maximum amount of time to be waited during a retry period.

name: Option<String>

The component’s name

penalty_duration: Option<String>

The amout of time that is used when the process penalizes a flowfile.

position: Option<Position>properties: HashMap<String, String>

The properties for the component. Properties whose value is not set will only contain the property name.

property_descriptors: HashMap<String, VersionedPropertyDescriptor>

The property descriptors for the component.

retried_relationships: Option<Vec<String>>

All the relationships should be retried.

retry_count: Option<i32>

Overall number of retries.

run_duration_millis: Option<i64>

The run duration for the processor in milliseconds.

scheduled_state: Option<VersionedProcessorScheduledState>

The scheduled state of the component

scheduling_period: Option<String>

The frequency with which to schedule the processor. The format of the value will depend on th value of schedulingStrategy.

scheduling_strategy: Option<String>

Indicates how the processor should be scheduled to run.

style: HashMap<String, String>

Stylistic data for rendering in a UI

type_: Option<String>

The type of the extension component

yield_duration: Option<String>

The amount of time that must elapse before this processor is scheduled again after yielding.

Implementations

impl VersionedProcessor

pub fn builder() -> VersionedProcessor