nifi_rs::proxy::v260

Module api

Modules

builder

Types for composing complex structures.

error

Error types.

Structs

AboutDto

AboutDto

AboutEntity

AboutEntity

AccessPolicyDto

AccessPolicyDto

AccessPolicyEntity

The access policies this user group belongs to. This field was incorrectly defined as an AccessPolicyEntity. For compatibility reasons the field will remain of this type, however only the fields that are present in the AccessPolicySummaryEntity will be populated here.

AccessPolicySummaryDto

AccessPolicySummaryDto

AccessPolicySummaryEntity

The access policies this user belongs to.

ActionDetailsDto

The details of the action.

ActionDto

ActionDto

ActionEntity

The actions.

ActivateControllerServicesEntity

ActivateControllerServicesEntity

AdditionalDetailsEntity

AdditionalDetailsEntity

AffectedComponentDto

AffectedComponentDto

AffectedComponentEntity

The set of all components in the flow that are referencing this Parameter

AllowableValueDto

AllowableValueDto

AllowableValueEntity

Allowable values for the property. If empty then the allowed values are not constrained.

AssetDto

The Asset.

AssetEntity

The asset entities

AssetReferenceDto

A list of identifiers of the assets that are referenced by the parameter

AssetsEntity

AssetsEntity

Attribute

The FlowFile attributes this processor writes/updates

AttributeDto

The attributes of the flowfile for the event.

AuthenticationConfigurationDto

AuthenticationConfigurationDto

AuthenticationConfigurationEntity

AuthenticationConfigurationEntity

BannerDto

BannerDto

BannerEntity

BannerEntity

BatchSettingsDto

The batch settings for data transmission.

BatchSize

The batch settings for data transmission.

BuildInfo

The build metadata for this component

BulletinBoardDto

BulletinBoardDto

BulletinBoardEntity

BulletinBoardEntity

BulletinBoardPatternParameter

BulletinBoardPatternParameter

BulletinDto

BulletinDto

BulletinEntity

The bulletins for this component.

Bundle

The details of the artifact that bundled this parameter provider.

BundleDto

If the property identifies a controller service this returns the bundle of the type, null otherwise.

ClientIdParameter

ClientIdParameter

ClusterDto

ClusterDto

ClusterEntity

ClusterEntity

ClusterSearchResultsEntity

ClusterSearchResultsEntity

ClusterSummaryDto

ClusterSummaryDto

ClusterSummaryEntity

ClusterSummaryEntity

ComponentDetailsDto

The details of the source component.

ComponentDifferenceDto

The list of differences for each component in the flow that is not the same between the two flows

ComponentHistoryDto

ComponentHistoryDto

ComponentHistoryEntity

ComponentHistoryEntity

ComponentManifest

The full specification of the bundle contents

ComponentReferenceDto

ComponentReferenceDto

ComponentReferenceEntity

Component this policy references if applicable.

ComponentRestrictionPermissionDto

Permissions for specific component restrictions.

ComponentSearchResultDto

The parameters that matched the search.

ComponentStateDto

The component state.

ComponentStateEntity

ComponentStateEntity

ComponentValidationResultDto

ComponentValidationResultDto

ComponentValidationResultEntity

A List of ComponentValidationResultEntity, one for each component that is validated

ComponentValidationResultsEntity

The Validation Results that were calculated for each component. This value may not be set until the request completes.

ConfigVerificationResultDto

The Results of the verification

ConfigurationAnalysisDto

The configuration analysis

ConfigurationAnalysisEntity

ConfigurationAnalysisEntity

ConnectableComponent

The destination of the connection.

ConnectableDto

The destination of the connection.

ConnectionDto

The connections in this flow snippet.

ConnectionEntity

The connections in this flow.

ConnectionStatisticsDto

ConnectionStatisticsDto

ConnectionStatisticsEntity

ConnectionStatisticsEntity

ConnectionStatisticsSnapshotDto

The connection status snapshot from the node.

ConnectionStatusDto

The status of the connection.

ConnectionStatusEntity

ConnectionStatusEntity

ConnectionStatusPredictionsSnapshotDto

Predictions, if available, for this connection (null if not available)

ConnectionStatusSnapshotDto

ConnectionStatusSnapshotDto

ConnectionStatusSnapshotEntity

The status of all connections in the process group.

ConnectionsEntity

ConnectionsEntity

ContentViewerDto

The Content Viewers.

ContentViewerEntity

ContentViewerEntity

ControllerBulletinsEntity

ControllerBulletinsEntity

ControllerConfigurationDto

The controller configuration.

ControllerConfigurationEntity

ControllerConfigurationEntity

ControllerDto

ControllerDto

ControllerEntity

ControllerEntity

ControllerServiceApi

Lists the APIs this Controller Service implements.

ControllerServiceApiDto

Lists the APIs this Controller Service implements.

ControllerServiceDefinition

Controller Services provided in this bundle

ControllerServiceDto

The controller services in this flow snippet.

ControllerServiceEntity

ControllerServiceEntity

ControllerServiceReferencingComponentDto

ControllerServiceReferencingComponentDto

ControllerServiceReferencingComponentEntity

All components referencing this controller service.

ControllerServiceReferencingComponentsEntity

ControllerServiceReferencingComponentsEntity

ControllerServiceRunStatusEntity

ControllerServiceRunStatusEntity

ControllerServiceStatusDto

The status for this ControllerService.

ControllerServiceTypesEntity

ControllerServiceTypesEntity

ControllerServicesEntity

ControllerServicesEntity

ControllerStatusDto

ControllerStatusDto

ControllerStatusEntity

ControllerStatusEntity

CopyRequestEntity

CopyRequestEntity

CopyResponseEntity

The response from copying.

CopySnippetRequestEntity

CopySnippetRequestEntity

CounterDto

CounterDto

CounterEntity

CounterEntity

CountersDto

CountersDto

CountersEntity

CountersEntity

CountersSnapshotDto

The counters from the node.

CreateActiveRequestEntity

CreateActiveRequestEntity

CurrentUserEntity

CurrentUserEntity

DateTimeParameter

DateTimeParameter

DefinedType

Indicates that this property is for selecting a controller service of the specified type

DifferenceDto

The differences in the component between the two flows

DimensionsDto

DimensionsDto

DocumentedTypeDto

DocumentedTypeDto

DropRequestDto

DropRequestDto

DropRequestEntity

DropRequestEntity

DynamicProperty

Describes the dynamic properties supported by this component

DynamicRelationship

If the processor supports dynamic relationships, this describes the dynamic relationship

ExplicitRestrictionDto

An optional collection of explicit restrictions. If specified, these explicit restrictions will be enfored.

ExternalControllerServiceReference

ExternalControllerServiceReference

FlowAnalysisResultEntity

FlowAnalysisResultEntity

FlowAnalysisRuleDefinition

Flow Analysis Rules provided in this bundle

FlowAnalysisRuleDto

FlowAnalysisRuleDto

FlowAnalysisRuleEntity

FlowAnalysisRuleEntity

FlowAnalysisRuleRunStatusEntity

FlowAnalysisRuleRunStatusEntity

FlowAnalysisRuleStatusDto

The status for this FlowAnalysisRule.

FlowAnalysisRuleTypesEntity

FlowAnalysisRuleTypesEntity

FlowAnalysisRuleViolationDto

FlowAnalysisRuleViolationDto

FlowAnalysisRulesEntity

FlowAnalysisRulesEntity

FlowBreadcrumbDto

This breadcrumb.

FlowBreadcrumbEntity

The breadcrumb of the process group.

FlowComparisonEntity

FlowComparisonEntity

FlowConfigurationDto

The controller configuration.

FlowConfigurationEntity

FlowConfigurationEntity

FlowDto

Flow containing the components that were created as part of this paste action.

FlowEntity

FlowEntity

FlowFileDto

FlowFileDto

FlowFileEntity

FlowFileEntity

FlowFileSummaryDto

The FlowFile summaries. The summaries will be populated once the request has completed.

FlowRegistryBranchDto

FlowRegistryBranchDto

FlowRegistryBranchEntity

FlowRegistryBranchEntity

FlowRegistryBranchesEntity

FlowRegistryBranchesEntity

FlowRegistryBucket

FlowRegistryBucket

FlowRegistryBucketDto

FlowRegistryBucketDto

FlowRegistryBucketEntity

FlowRegistryBucketEntity

FlowRegistryBucketsEntity

FlowRegistryBucketsEntity

FlowRegistryClientDto

FlowRegistryClientDto

FlowRegistryClientEntity

FlowRegistryClientEntity

FlowRegistryClientTypesEntity

FlowRegistryClientTypesEntity

FlowRegistryClientsEntity

FlowRegistryClientsEntity

FlowRegistryPermissions

FlowRegistryPermissions

FlowSnippetDto

The contents of this process group.

FunnelDto

The funnels in this flow snippet.

FunnelEntity

The funnels in this flow.

FunnelsEntity

FunnelsEntity

GarbageCollectionDto

The garbage collection details.

HistoryDto

HistoryDto

HistoryEntity

HistoryEntity

InputPortsEntity

InputPortsEntity

IntegerParameter

IntegerParameter

JmxMetricsResultDto

JmxMetricsResultDto

JmxMetricsResultsEntity

JmxMetricsResultsEntity

LabelDto

The labels in this flow snippet.

LabelEntity

The labels in this flow.

LabelsEntity

LabelsEntity

LatestProvenanceEventsDto

LatestProvenanceEventsDto

LatestProvenanceEventsEntity

LatestProvenanceEventsEntity

LineageDto

LineageDto

LineageEntity

LineageEntity

LineageRequestDto

The initial lineage result.

LineageResultsDto

The results of the lineage query.

ListingRequestDto

ListingRequestDto

ListingRequestEntity

ListingRequestEntity

LongParameter

LongParameter

MultiProcessorUseCase

A list of use cases that have been documented that involve this Processor in conjunction with other Processors

NarCoordinateDto

The coordinate of another NAR that the this NAR is dependent on, or null if not dependent on another NAR.

NarDetailsEntity

NarDetailsEntity

NarSummariesEntity

NarSummariesEntity

NarSummaryDto

The NAR summary

NarSummaryEntity

NarSummaryEntity

NodeConnectionStatisticsSnapshotDto

A list of status snapshots for each node

NodeConnectionStatusSnapshotDto

A list of status snapshots for each node

NodeCountersSnapshotDto

A Counters snapshot for each node in the cluster. If the NiFi instance is a standalone instance, rather than a cluster, this may be null.

NodeDto

NodeDto

NodeEntity

NodeEntity

NodeEventDto

The node’s events.

NodePortStatusSnapshotDto

A status snapshot for each node in the cluster. If the NiFi instance is a standalone instance, rather than a cluster, this may be null.

NodeProcessGroupStatusSnapshotDto

The status reported by each node in the cluster. If the NiFi instance is a standalone instance, rather than a clustered instance, this value may be null.

NodeProcessorStatusSnapshotDto

A status snapshot for each node in the cluster. If the NiFi instance is a standalone instance, rather than a cluster, this may be null.

NodeRemoteProcessGroupStatusSnapshotDto

A status snapshot for each node in the cluster. If the NiFi instance is a standalone instance, rather than a cluster, this may be null.

NodeReplayLastEventSnapshotDto

The node-wise results

NodeSearchResultDto

NodeSearchResultDto

NodeStatusSnapshotsDto

The NodeStatusSnapshotsDTO objects that provide the actual metric values for the component, for each node. If the NiFi instance is not clustered, this value will be null.

NodeSystemDiagnosticsSnapshotDto

A systems diagnostics snapshot for each node in the cluster. If the NiFi instance is a standalone instance, rather than a cluster, this may be null.

OutputPortsEntity

OutputPortsEntity

ParameterContextDto

The Parameter Context that is being operated on. This may not be populated until the request has successfully completed.

ParameterContextEntity

ParameterContextEntity

ParameterContextReferenceDto

ParameterContextReferenceDto

ParameterContextReferenceEntity

The Parameter Context, or null if no Parameter Context has been bound to the Process Group

ParameterContextUpdateEntity

The Parameter Contexts updated by this Parameter Provider. This may not be populated until the request has successfully completed.

ParameterContextUpdateRequestDto

The Update Request

ParameterContextUpdateRequestEntity

ParameterContextUpdateRequestEntity

ParameterContextUpdateStepDto

The steps that are required in order to complete the request, along with the status of each

ParameterContextValidationRequestDto

The Update Request

ParameterContextValidationRequestEntity

ParameterContextValidationRequestEntity

ParameterContextValidationStepDto

The steps that are required in order to complete the request, along with the status of each

ParameterContextsEntity

ParameterContextsEntity

ParameterDto

The parameter information

ParameterEntity

The name of the Parameter

ParameterGroupConfigurationEntity

Configuration for any fetched parameter groups.

ParameterProviderApplyParametersRequestDto

The Apply Parameters Request

ParameterProviderApplyParametersRequestEntity

ParameterProviderApplyParametersRequestEntity

ParameterProviderApplyParametersUpdateStepDto

The steps that are required in order to complete the request, along with the status of each

ParameterProviderConfigurationDto

ParameterProviderConfigurationDto

ParameterProviderConfigurationEntity

Optional configuration for a Parameter Provider

ParameterProviderDefinition

Parameter Providers provided in this bundle

ParameterProviderDto

ParameterProviderDto

ParameterProviderEntity

ParameterProviderEntity

ParameterProviderParameterApplicationEntity

ParameterProviderParameterApplicationEntity

ParameterProviderParameterFetchEntity

ParameterProviderParameterFetchEntity

ParameterProviderReference

ParameterProviderReference

ParameterProviderReferencingComponentDto

ParameterProviderReferencingComponentDto

ParameterProviderReferencingComponentEntity

The Parameter Contexts that reference this Parameter Provider

ParameterProviderReferencingComponentsEntity

ParameterProviderReferencingComponentsEntity

ParameterProviderTypesEntity

ParameterProviderTypesEntity

ParameterProvidersEntity

ParameterProvidersEntity

ParameterStatusDto

The status of all provided parameters for this parameter provider

PasteRequestEntity

PasteRequestEntity

PasteResponseEntity

PasteResponseEntity

PeerDto

PeerDto

PeersEntity

PeersEntity

PermissionsDto

The permissions for this component.

PortDto

The output ports available to received data from the NiFi.

PortEntity

The output ports in this flow.

PortRunStatusEntity

PortRunStatusEntity

PortStatusDto

The status of the port.

PortStatusEntity

PortStatusEntity

PortStatusSnapshotDto

PortStatusSnapshotDto

PortStatusSnapshotEntity

The status of all output ports in the process group.

Position

The position of a component on the graph

PositionDto

The position of this component in the UI if applicable.

PreviousValueDto

Previous values for a given property.

PrioritizerTypesEntity

PrioritizerTypesEntity

ProcessGroupDto

ProcessGroupDto

ProcessGroupEntity

ProcessGroupEntity

ProcessGroupFlowDto

ProcessGroupFlowDto

ProcessGroupFlowEntity

ProcessGroupFlowEntity

ProcessGroupImportEntity

ProcessGroupImportEntity

ProcessGroupNameDto

The Process Group that the component belongs to

ProcessGroupReplaceRequestDto

The Process Group Change Request

ProcessGroupReplaceRequestEntity

ProcessGroupReplaceRequestEntity

ProcessGroupStatusDto

The status of the process group.

ProcessGroupStatusEntity

ProcessGroupStatusEntity

ProcessGroupStatusSnapshotDto

The process group status snapshot from the node.

ProcessGroupStatusSnapshotEntity

The status of all process groups in the process group.

ProcessGroupUploadEntity

ProcessGroupUploadEntity

ProcessGroupsEntity

ProcessGroupsEntity

ProcessingPerformanceStatusDto

Represents the processor’s processing performance.

ProcessorConfigDto

The configuration details for the processor. These details will be included in a response if the verbose flag is included in a request.

ProcessorConfiguration

A description of how to configure the Processor to perform the task described in the use case

ProcessorDefinition

Processors provided in this bundle

ProcessorDto

ProcessorDto

ProcessorEntity

ProcessorEntity

ProcessorRunStatusDetailsDto

The details of a Processor’s run status

ProcessorRunStatusDetailsEntity

ProcessorRunStatusDetailsEntity

ProcessorRunStatusEntity

ProcessorRunStatusEntity

ProcessorStatusDto

ProcessorStatusDto

ProcessorStatusEntity

ProcessorStatusEntity

ProcessorStatusSnapshotDto

The processor status snapshot from the node.

ProcessorStatusSnapshotEntity

The status of all processors in the process group.

ProcessorTypesEntity

ProcessorTypesEntity

ProcessorsEntity

ProcessorsEntity

ProcessorsRunStatusDetailsEntity

ProcessorsRunStatusDetailsEntity

PropertyAllowableValue

A list of the allowable values for the property

PropertyDependency

The dependencies that this property has on other properties

PropertyDependencyDto

A list of dependencies that must be met in order for this Property to be relevant. If any of these dependencies is not met, the property described by this Property Descriptor is not relevant.

PropertyDescriptor

Descriptions of configuration properties applicable to this component.

PropertyDescriptorDto

The descriptors for the reporting tasks properties.

PropertyDescriptorEntity

PropertyDescriptorEntity

PropertyHistoryDto

The history for the properties of the component.

PropertyResourceDefinition

Indicates that this property references external resources

ProvenanceDto

ProvenanceDto

ProvenanceEntity

ProvenanceEntity

ProvenanceEventDto

The provenance events that matched the search criteria.

ProvenanceEventEntity

ProvenanceEventEntity

ProvenanceLinkDto

The links between the nodes in the lineage.

ProvenanceNodeDto

The nodes in the lineage.

ProvenanceOptionsDto

ProvenanceOptionsDto

ProvenanceOptionsEntity

ProvenanceOptionsEntity

ProvenanceRequestDto

The provenance request.

ProvenanceResultsDto

The provenance results.

ProvenanceSearchValueDto

The search terms used to perform the search.

ProvenanceSearchableFieldDto

The available searchable field for the NiFi.

QueueSizeDto

The size of the queue

RegisteredFlow

RegisteredFlow

RegisteredFlowSnapshot

RegisteredFlowSnapshot

RegisteredFlowSnapshotMetadata

RegisteredFlowSnapshotMetadata

RegisteredFlowVersionInfo

RegisteredFlowVersionInfo

Relationship

The supported relationships for this processor.

RelationshipDto

The available relationships that the processor currently supports.

RemotePortRunStatusEntity

RemotePortRunStatusEntity

RemoteProcessGroupContentsDto

The contents of the remote process group. Will contain available input/output ports.

RemoteProcessGroupDto

RemoteProcessGroupDto

RemoteProcessGroupEntity

RemoteProcessGroupEntity

RemoteProcessGroupPortDto

The output ports from which data can be retrieved.

RemoteProcessGroupPortEntity

RemoteProcessGroupPortEntity

RemoteProcessGroupStatusDto

The status of the remote process group.

RemoteProcessGroupStatusEntity

RemoteProcessGroupStatusEntity

RemoteProcessGroupStatusSnapshotDto

The remote process group status snapshot from the node.

RemoteProcessGroupStatusSnapshotEntity

The status of all remote process groups in the process group.

RemoteProcessGroupsEntity

RemoteProcessGroupsEntity

ReplayLastEventRequestEntity

ReplayLastEventRequestEntity

ReplayLastEventResponseEntity

ReplayLastEventResponseEntity

ReplayLastEventSnapshotDto

The snapshot from the node

ReportingTaskDefinition

Reporting Tasks provided in this bundle

ReportingTaskDto

ReportingTaskDto

ReportingTaskEntity

ReportingTaskEntity

ReportingTaskRunStatusEntity

ReportingTaskRunStatusEntity

ReportingTaskStatusDto

The status for this ReportingTask.

ReportingTaskTypesEntity

ReportingTaskTypesEntity

ReportingTasksEntity

ReportingTasksEntity

RequiredPermissionDto

The required permission necessary for this restriction.

ResourceClaimDetailsDto

ResourceClaimDetailsDto

ResourceDto

ResourceDto

ResourcesEntity

ResourcesEntity

Restriction

Explicit restrictions that indicate a require permission to use the component

RevisionDto

The revision of the Process Group

RunStatusDetailsRequestEntity

RunStatusDetailsRequestEntity

RuntimeManifest

RuntimeManifest

RuntimeManifestEntity

RuntimeManifestEntity

ScheduleComponentsEntity

ScheduleComponentsEntity

SchedulingDefaults

Scheduling defaults for components defined in this manifest

SearchResultGroupDto

The nearest versioned ancestor group of the component that matched the search.

SearchResultsDto

SearchResultsDto

SearchResultsEntity

SearchResultsEntity

SnippetDto

The snippet.

SnippetEntity

SnippetEntity

StartVersionControlRequestEntity

StartVersionControlRequestEntity

StateEntryDto

The state.

StateMapDto

The local state for this component.

Stateful

Indicates if the component stores state

StatusDescriptorDto

The Descriptors that provide information on each of the metrics provided in the status history

StatusHistoryDto

StatusHistoryDto

StatusHistoryEntity

StatusHistoryEntity

StatusSnapshotDto

A list of StatusSnapshotDTO objects that provide the actual metric values for the component for this node.

StorageUsageDto

The provenance repository storage usage.

StreamingOutput

StreamingOutput

SubmitReplayRequestEntity

SubmitReplayRequestEntity

SupportedMimeTypesDto

The mime types this Content Viewer supports.

SystemDiagnosticsDto

SystemDiagnosticsDto

SystemDiagnosticsEntity

SystemDiagnosticsEntity

SystemDiagnosticsSnapshotDto

The System Diagnostics snapshot from the node.

SystemResourceConsideration

The system resource considerations for the given component

TenantDto

TenantDto

TenantEntity

The set of user group IDs associated with this access policy.

TenantsEntity

TenantsEntity

TransactionResultEntity

TransactionResultEntity

UpdateControllerServiceReferenceRequestEntity

UpdateControllerServiceReferenceRequestEntity

UseCase

A list of use cases that have been documented for this Processor

UserDto

UserDto

UserEntity

UserEntity

UserGroupDto

UserGroupDto

UserGroupEntity

UserGroupEntity

UserGroupsEntity

UserGroupsEntity

UsersEntity

UsersEntity

VerifyConfigRequestDto

The request

VerifyConfigRequestEntity

VerifyConfigRequestEntity

VerifyConfigUpdateStepDto

The steps that are required in order to complete the request, along with the status of each

VersionControlComponentMappingEntity

VersionControlComponentMappingEntity

VersionControlInformationDto

The Version Control information

VersionControlInformationEntity

VersionControlInformationEntity

VersionInfoDto

The nifi, os, java, and build version information

VersionedAsset

The assets that are referenced by this parameter

VersionedConnection

The Connections

VersionedControllerService

The Controller Services

VersionedFlowCoordinates

The coordinates where the remote flow is stored, or null if the Process Group is not directly under Version Control

VersionedFlowDto

The versioned flow

VersionedFlowEntity

VersionedFlowEntity

VersionedFlowSnapshotEntity

VersionedFlowSnapshotEntity

VersionedFlowSnapshotMetadataEntity

VersionedFlowSnapshotMetadataEntity

VersionedFlowSnapshotMetadataSetEntity

VersionedFlowSnapshotMetadataSetEntity

VersionedFlowUpdateRequestDto

The Flow Update Request

VersionedFlowUpdateRequestEntity

VersionedFlowUpdateRequestEntity

VersionedFlowsEntity

VersionedFlowsEntity

VersionedFunnel

The Funnels

VersionedLabel

The Labels

VersionedParameter

The parameters in the context

VersionedParameterContext

VersionedParameterContext

VersionedPort

The Output Ports

VersionedProcessGroup

VersionedProcessGroup

VersionedProcessor

The Processors

VersionedPropertyDescriptor

The property descriptors for the component.

VersionedRemoteGroupPort

A Set of Output Ports that can be connected to, in order to pull data from the remote NiFi instance

VersionedRemoteProcessGroup

The Remote Process Groups

VersionedReportingTask

The reporting tasks

VersionedReportingTaskImportRequestEntity

VersionedReportingTaskImportRequestEntity

VersionedReportingTaskImportResponseEntity

VersionedReportingTaskImportResponseEntity

VersionedReportingTaskSnapshot

VersionedReportingTaskSnapshot

VersionedResourceDefinition

Returns the Resource Definition that defines which type(s) of resource(s) this property references, if any

Enums

AccessPolicyDtoAction

The action associated with this access policy.

AccessPolicySummaryDtoAction

The action associated with this access policy.

ActivateControllerServicesEntityState

The desired state of the descendant components

AffectedComponentDtoReferenceType

The type of this component

AffectedComponentEntityReferenceType

The type of component referenced

ComponentValidationResultDtoReferenceType

The type of this component

ConfigVerificationResultDtoOutcome

The outcome of the verification

ConnectableComponentType

The type of component the connectable is.

ConnectableDtoType

The type of component the connectable is.

ConnectionEntityDestinationType

The type of component the destination connectable is.

ConnectionEntitySourceType

The type of component the source connectable is.

ConnectionStatusSnapshotDtoLoadBalanceStatus

The load balance status of the connection

ControllerServiceDtoState

The state of the controller service.

ControllerServiceDtoValidationStatus

Indicates whether the ControllerService is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the ControllerService is valid)

ControllerServiceReferencingComponentDtoReferenceType

The type of reference this is.

ControllerServiceRunStatusEntityState

The run status of the ControllerService.

ControllerServiceStatusDtoRunStatus

The run status of this ControllerService

ControllerServiceStatusDtoValidationStatus

Indicates whether the component is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the component is valid)

DynamicPropertyExpressionLanguageScope

The scope of the expression language support

FlowAnalysisRuleDtoState

The state of the flow analysis rule.

FlowAnalysisRuleDtoValidationStatus

Indicates whether the Flow Analysis Rule is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Flow Analysis Rule is valid)

FlowAnalysisRuleRunStatusEntityState

The state of the FlowAnalysisRule.

FlowAnalysisRuleStatusDtoRunStatus

The run status of this FlowAnalysisRule

FlowAnalysisRuleStatusDtoValidationStatus

Indicates whether the component is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the component is valid)

FlowBreadcrumbEntityVersionedFlowState

The current state of the Process Group, as it relates to the Versioned Flow

FlowRegistryClientDtoValidationStatus

Indicates whether the Registry Client is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Registry Client is valid)

LineageRequestDtoLineageRequestType

The type of lineage request. PARENTS will return the lineage for the flowfiles that are parents of the specified event. CHILDREN will return the lineage for the flowfiles that are children of the specified event. FLOWFILE will return the lineage for the specified flowfile.

ParameterGroupConfigurationEntityParameterSensitivitiesValue

All fetched parameter names that should be applied.

ParameterProviderDtoValidationStatus

Indicates whether the Parameter Provider is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Parameter Provider is valid)

ParameterStatusDtoStatus

Indicates the status of the parameter, compared to the existing parameter context

PortDtoPortFunction

Specifies how the Port functions

PortDtoState

The state of the port.

PortDtoType

The type of port.

PortRunStatusEntityState

The run status of the Port.

PortStatusDtoRunStatus

The run status of the port.

PortStatusSnapshotDtoRunStatus

The run status of the port.

ProcessGroupDtoExecutionEngine

The Execution Engine that should be used to run the flow represented by this Process Group.

ProcessGroupDtoFlowfileConcurrency

The FlowFile Concurrency for this Process Group.

ProcessGroupDtoFlowfileOutboundPolicy

The Outbound Policy that is used for determining how FlowFiles should be transferred out of the Process Group.

ProcessGroupDtoStatelessGroupScheduledState

If the Process Group is configured to run in using the Stateless Engine, represents the current state. Otherwise, will be STOPPED.

ProcessGroupEntityProcessGroupUpdateStrategy

Determines the process group update strategy

ProcessGroupEntityVersionedFlowState

The current state of the Process Group, as it relates to the Versioned Flow

ProcessGroupStatusSnapshotDtoVersionedFlowState

The current state of the Process Group, as it relates to the Versioned Flow

ProcessorDefinitionInputRequirement

Any input requirements this processor has.

ProcessorDtoState

The state of the processor

ProcessorDtoValidationStatus

Indicates whether the Processor is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Processor is valid)

ProcessorRunStatusDetailsDtoRunStatus

The run status of the processor

ProcessorRunStatusEntityState

The run status of the Processor.

ProcessorStatusDtoRunStatus

The run status of the Processor

ProcessorStatusSnapshotDtoExecutionNode

Indicates the node where the process will execute.

ProcessorStatusSnapshotDtoRunStatus

The state of the processor.

PropertyDescriptorExpressionLanguageScope

The scope of expression language supported by this property

PropertyResourceDefinitionCardinality

The cardinality of the resource definition (i.e. single or multiple)

PropertyResourceDefinitionResourceTypesItem

The types of resources that can be referenced

ProvenanceNodeDtoType

The type of the node.

RemotePortRunStatusEntityState

The run status of the RemotePort.

RemoteProcessGroupStatusDtoValidationStatus

Indicates whether the component is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the component is valid)

ReplayLastEventRequestEntityNodes

Which nodes are to replay their last provenance event.

ReplayLastEventResponseEntityNodes

Which nodes were requested to replay their last provenance event.

ReportingTaskDtoState

The state of the reporting task.

ReportingTaskDtoValidationStatus

Indicates whether the Reporting Task is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the Reporting Task is valid)

ReportingTaskRunStatusEntityState

The run status of the ReportingTask.

ReportingTaskStatusDtoRunStatus

The run status of this ReportingTask

ReportingTaskStatusDtoValidationStatus

Indicates whether the component is valid, invalid, or still in the process of validating (i.e., it is unknown whether or not the component is valid)

ScheduleComponentsEntityState

The desired state of the descendant components

SchedulingDefaultsDefaultSchedulingStrategy

The name of the default scheduling strategy

StatefulScopesItem

Indicates the Scope(s) associated with the State that is stored and retrieved

UpdateControllerServiceReferenceRequestEntityState

The new state of the references for the controller service.

UseCaseInputRequirement

Specifies whether an incoming FlowFile is expected for this use case

VersionControlInformationDtoState

The current state of the Process Group, as it relates to the Versioned Flow

VersionedConnectionComponentType

VersionedConnectionComponentType

VersionedControllerServiceComponentType

VersionedControllerServiceComponentType

VersionedControllerServiceScheduledState

The ScheduledState denoting whether the Controller Service is ENABLED or DISABLED

VersionedFlowDtoAction

The action being performed

VersionedFunnelComponentType

VersionedFunnelComponentType

VersionedLabelComponentType

VersionedLabelComponentType

VersionedParameterContextComponentType

VersionedParameterContextComponentType

VersionedPortComponentType

VersionedPortComponentType

VersionedPortPortFunction

Specifies how the Port should function

VersionedPortScheduledState

The scheduled state of the component

VersionedPortType

The type of port.

VersionedProcessGroupComponentType

VersionedProcessGroupComponentType

VersionedProcessGroupExecutionEngine

The Execution Engine that should be used to run the components within the group.

VersionedProcessGroupScheduledState

The Scheduled State of the Process Group, if the group is configured to use the Stateless Execution Engine. Otherwise, this value has no relevance.

VersionedProcessorComponentType

VersionedProcessorComponentType

VersionedProcessorScheduledState

The scheduled state of the component

VersionedRemoteGroupPortComponentType

VersionedRemoteGroupPortComponentType

VersionedRemoteGroupPortScheduledState

The scheduled state of the component

VersionedRemoteProcessGroupComponentType

VersionedRemoteProcessGroupComponentType

VersionedReportingTaskComponentType

VersionedReportingTaskComponentType

VersionedReportingTaskScheduledState

Indicates the scheduled state for the Reporting Task

VersionedResourceDefinitionCardinality

The cardinality of the resource

VersionedResourceDefinitionResourceTypesItem

The types of resource that the Property Descriptor is allowed to reference