nifi_rs::proxy::v260::api

Struct ControllerServiceReferencingComponentDto

pub struct ControllerServiceReferencingComponentDto {
    pub active_thread_count: Option<i32>,
    pub descriptors: HashMap<String, PropertyDescriptorDto>,
    pub group_id: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
    pub properties: HashMap<String, String>,
    pub reference_cycle: Option<bool>,
    pub reference_type: Option<ControllerServiceReferencingComponentDtoReferenceType>,
    pub referencing_components: Option<Vec<ControllerServiceReferencingComponentEntity>>,
    pub state: Option<String>,
    pub type_: Option<String>,
    pub validation_errors: Vec<String>,
}

ControllerServiceReferencingComponentDto

JSON schema

{
 "type": "object",
 "properties": {
   "activeThreadCount": {
     "description": "The number of active threads for the referencing component.",
     "type": "integer",
     "format": "int32"
   },
   "descriptors": {
     "description": "The descriptors for the component properties.",
     "type": "object",
     "additionalProperties": {
       "$ref": "#/components/schemas/PropertyDescriptorDTO"
     }
   },
   "groupId": {
     "description": "The group id for the component referencing a controller service. If this component is another controller service or a reporting task, this field is blank.",
     "type": "string"
   },
   "id": {
     "description": "The id of the component referencing a controller service.",
     "type": "string"
   },
   "name": {
     "description": "The name of the component referencing a controller service.",
     "type": "string"
   },
   "properties": {
     "description": "The properties for the component.",
     "type": "object",
     "additionalProperties": {
       "description": "The properties for the component.",
       "type": "string"
     }
   },
   "referenceCycle": {
     "description": "If the referencing component represents a controller service, this indicates whether it has already been represented in this hierarchy.",
     "type": "boolean"
   },
   "referenceType": {
     "description": "The type of reference this is.",
     "type": "string",
     "enum": [
       "Processor",
       "ControllerService",
       "ReportingTask",
       "FlowRegistryClient"
     ]
   },
   "referencingComponents": {
     "description": "If the referencing component represents a controller service, these are the components that reference it.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/ControllerServiceReferencingComponentEntity"
     },
     "uniqueItems": true
   },
   "state": {
     "description": "The scheduled state of a processor or reporting task referencing a controller service. If this component is another controller service, this field represents the controller service state.",
     "type": "string"
   },
   "type": {
     "description": "The type of the component referencing a controller service in simple Java class name format without package name.",
     "type": "string"
   },
   "validationErrors": {
     "description": "The validation errors for the component.",
     "type": "array",
     "items": {
       "description": "The validation errors for the component.",
       "type": "string"
     }
   }
 }
}

Fields
active_thread_count: Option<i32>

The number of active threads for the referencing component.

descriptors: HashMap<String, PropertyDescriptorDto>

The descriptors for the component properties.

group_id: Option<String>

The group id for the component referencing a controller service. If this component is another controller service or a reporting task, this field is blank.

id: Option<String>

The id of the component referencing a controller service.

name: Option<String>

The name of the component referencing a controller service.

properties: HashMap<String, String>

The properties for the component.

reference_cycle: Option<bool>

If the referencing component represents a controller service, this indicates whether it has already been represented in this hierarchy.

reference_type: Option<ControllerServiceReferencingComponentDtoReferenceType>

The type of reference this is.

referencing_components: Option<Vec<ControllerServiceReferencingComponentEntity>>

If the referencing component represents a controller service, these are the components that reference it.

state: Option<String>

The scheduled state of a processor or reporting task referencing a controller service. If this component is another controller service, this field represents the controller service state.

type_: Option<String>

The type of the component referencing a controller service in simple Java class name format without package name.

validation_errors: Vec<String>

The validation errors for the component.

Implementations

impl ControllerServiceReferencingComponentDto

pub fn builder() -> ControllerServiceReferencingComponentDto