nifi_rs::proxy::v260::api

Struct ProcessGroupDto

pub struct ProcessGroupDto {

    pub active_remote_port_count: Option<i32>,
    pub comments: Option<String>,
    pub contents: Option<FlowSnippetDto>,
    pub default_back_pressure_data_size_threshold: Option<String>,
    pub default_back_pressure_object_threshold: Option<i64>,
    pub default_flow_file_expiration: Option<String>,
    pub disabled_count: Option<i32>,
    pub execution_engine: Option<ProcessGroupDtoExecutionEngine>,
    pub flowfile_concurrency: Option<ProcessGroupDtoFlowfileConcurrency>,
    pub flowfile_outbound_policy: Option<ProcessGroupDtoFlowfileOutboundPolicy>,
    pub id: Option<String>,
    pub inactive_remote_port_count: Option<i32>,
    pub input_port_count: Option<i32>,
    pub invalid_count: Option<i32>,
    pub local_input_port_count: Option<i32>,
    pub local_output_port_count: Option<i32>,
    pub locally_modified_and_stale_count: Option<i32>,
    pub locally_modified_count: Option<i32>,
    pub log_file_suffix: Option<String>,
    pub max_concurrent_tasks: Option<i32>,
    pub name: Option<String>,
    pub output_port_count: Option<i32>,
    pub parameter_context: Option<ParameterContextReferenceEntity>,
    pub parent_group_id: Option<String>,
    pub position: Option<PositionDto>,
    pub public_input_port_count: Option<i32>,
    pub public_output_port_count: Option<i32>,
    pub running_count: Option<i32>,
    pub stale_count: Option<i32>,
    pub stateless_flow_timeout: Option<String>,
    pub stateless_group_scheduled_state: Option<ProcessGroupDtoStatelessGroupScheduledState>,
    pub stopped_count: Option<i32>,
    pub sync_failure_count: Option<i32>,
    pub up_to_date_count: Option<i32>,
    pub version_control_information: Option<VersionControlInformationDto>,
    pub versioned_component_id: Option<String>,

}

ProcessGroupDto

JSON schema

{
 "type": "object",
 "properties": {
   "activeRemotePortCount": {
     "description": "The number of active remote ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "comments": {
     "description": "The comments for the process group.",
     "type": "string"
   },
   "contents": {
     "$ref": "#/components/schemas/FlowSnippetDTO"
   },
   "defaultBackPressureDataSizeThreshold": {
     "description": "Default value used in this Process Group for the maximum data size of objects that can be queued before back pressure is applied.",
     "type": "string"
   },
   "defaultBackPressureObjectThreshold": {
     "description": "Default value used in this Process Group for the maximum number of objects that can be queued before back pressure is applied.",
     "type": "integer",
     "format": "int64"
   },
   "defaultFlowFileExpiration": {
     "description": "The default FlowFile Expiration for this Process Group.",
     "type": "string"
   },
   "disabledCount": {
     "description": "The number of disabled components in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "executionEngine": {
     "description": "The Execution Engine that should be used to run the flow represented by this Process Group.",
     "type": "string",
     "enum": [
       "STATELESS",
       "STANDARD",
       "INHERITED"
     ]
   },
   "flowfileConcurrency": {
     "description": "The FlowFile Concurrency for this Process Group.",
     "type": "string",
     "enum": [
       "UNBOUNDED",
       "SINGLE_FLOWFILE_PER_NODE",
       "SINGLE_BATCH_PER_NODE"
     ]
   },
   "flowfileOutboundPolicy": {
     "description": "The Outbound Policy that is used for determining how FlowFiles should be transferred out of the Process Group.",
     "type": "string",
     "enum": [
       "STREAM_WHEN_AVAILABLE",
       "BATCH_OUTPUT"
     ]
   },
   "id": {
     "description": "The id of the component.",
     "type": "string"
   },
   "inactiveRemotePortCount": {
     "description": "The number of inactive remote ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "inputPortCount": {
     "description": "The number of input ports in the process group.",
     "readOnly": true,
     "type": "integer",
     "format": "int32"
   },
   "invalidCount": {
     "description": "The number of invalid components in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "localInputPortCount": {
     "description": "The number of local input ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "localOutputPortCount": {
     "description": "The number of local output ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "locallyModifiedAndStaleCount": {
     "description": "The number of locally modified and stale versioned process groups in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "locallyModifiedCount": {
     "description": "The number of locally modified versioned process groups in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "logFileSuffix": {
     "description": "The log file suffix for this Process Group for dedicated logging.",
     "type": "string"
   },
   "maxConcurrentTasks": {
     "description": "The maximum number of concurrent tasks to use when running the flow using the Stateless Engine",
     "type": "integer",
     "format": "int32"
   },
   "name": {
     "description": "The name of the process group.",
     "type": "string"
   },
   "outputPortCount": {
     "description": "The number of output ports in the process group.",
     "readOnly": true,
     "type": "integer",
     "format": "int32"
   },
   "parameterContext": {
     "$ref": "#/components/schemas/ParameterContextReferenceEntity"
   },
   "parentGroupId": {
     "description": "The id of parent process group of this component if applicable.",
     "type": "string"
   },
   "position": {
     "$ref": "#/components/schemas/PositionDTO"
   },
   "publicInputPortCount": {
     "description": "The number of public input ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "publicOutputPortCount": {
     "description": "The number of public output ports in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "runningCount": {
     "description": "The number of running components in this process group.",
     "type": "integer",
     "format": "int32"
   },
   "staleCount": {
     "description": "The number of stale versioned process groups in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "statelessFlowTimeout": {
     "description": "The maximum amount of time that the flow can be run using the Stateless Engine before the flow times out",
     "type": "string"
   },
   "statelessGroupScheduledState": {
     "description": "If the Process Group is configured to run in using the Stateless Engine, represents the current state. Otherwise, will be STOPPED.",
     "type": "string",
     "enum": [
       "STOPPED",
       "RUNNING"
     ]
   },
   "stoppedCount": {
     "description": "The number of stopped components in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "syncFailureCount": {
     "description": "The number of versioned process groups in the process group that are unable to sync to a registry.",
     "type": "integer",
     "format": "int32"
   },
   "upToDateCount": {
     "description": "The number of up to date versioned process groups in the process group.",
     "type": "integer",
     "format": "int32"
   },
   "versionControlInformation": {
     "$ref": "#/components/schemas/VersionControlInformationDTO"
   },
   "versionedComponentId": {
     "description": "The ID of the corresponding component that is under version control",
     "type": "string"
   }
 }
}

Fields
active_remote_port_count: Option<i32>

The number of active remote ports in the process group.

comments: Option<String>

The comments for the process group.

contents: Option<FlowSnippetDto>default_back_pressure_data_size_threshold: Option<String>

Default value used in this Process Group for the maximum data size of objects that can be queued before back pressure is applied.

default_back_pressure_object_threshold: Option<i64>

Default value used in this Process Group for the maximum number of objects that can be queued before back pressure is applied.

default_flow_file_expiration: Option<String>

The default FlowFile Expiration for this Process Group.

disabled_count: Option<i32>

The number of disabled components in the process group.

execution_engine: Option<ProcessGroupDtoExecutionEngine>

The Execution Engine that should be used to run the flow represented by this Process Group.

flowfile_concurrency: Option<ProcessGroupDtoFlowfileConcurrency>

The FlowFile Concurrency for this Process Group.

flowfile_outbound_policy: Option<ProcessGroupDtoFlowfileOutboundPolicy>

The Outbound Policy that is used for determining how FlowFiles should be transferred out of the Process Group.

id: Option<String>

The id of the component.

inactive_remote_port_count: Option<i32>

The number of inactive remote ports in the process group.

input_port_count: Option<i32>

The number of input ports in the process group.

invalid_count: Option<i32>

The number of invalid components in the process group.

local_input_port_count: Option<i32>

The number of local input ports in the process group.

local_output_port_count: Option<i32>

The number of local output ports in the process group.

locally_modified_and_stale_count: Option<i32>

The number of locally modified and stale versioned process groups in the process group.

locally_modified_count: Option<i32>

The number of locally modified versioned process groups in the process group.

log_file_suffix: Option<String>

The log file suffix for this Process Group for dedicated logging.

max_concurrent_tasks: Option<i32>

The maximum number of concurrent tasks to use when running the flow using the Stateless Engine

name: Option<String>

The name of the process group.

output_port_count: Option<i32>

The number of output ports in the process group.

parameter_context: Option<ParameterContextReferenceEntity>parent_group_id: Option<String>

The id of parent process group of this component if applicable.

position: Option<PositionDto>public_input_port_count: Option<i32>

The number of public input ports in the process group.

public_output_port_count: Option<i32>

The number of public output ports in the process group.

running_count: Option<i32>

The number of running components in this process group.

stale_count: Option<i32>

The number of stale versioned process groups in the process group.

stateless_flow_timeout: Option<String>

The maximum amount of time that the flow can be run using the Stateless Engine before the flow times out

stateless_group_scheduled_state: Option<ProcessGroupDtoStatelessGroupScheduledState>

If the Process Group is configured to run in using the Stateless Engine, represents the current state. Otherwise, will be STOPPED.

stopped_count: Option<i32>

The number of stopped components in the process group.

sync_failure_count: Option<i32>

The number of versioned process groups in the process group that are unable to sync to a registry.

up_to_date_count: Option<i32>

The number of up to date versioned process groups in the process group.

version_control_information: Option<VersionControlInformationDto>versioned_component_id: Option<String>

The ID of the corresponding component that is under version control

Implementations

impl ProcessGroupDto

pub fn builder() -> ProcessGroupDto