nifi_rs::proxy::v260::api

Struct ProcessorDefinition

pub struct ProcessorDefinition {

    pub additional_details: Option<bool>,
    pub artifact: Option<String>,
    pub build_info: Option<BuildInfo>,
    pub default_bulletin_level: Option<String>,
    pub default_concurrent_tasks_by_scheduling_strategy: HashMap<String, i32>,
    pub default_penalty_duration: Option<String>,
    pub default_scheduling_period_by_scheduling_strategy: HashMap<String, String>,
    pub default_scheduling_strategy: Option<String>,
    pub default_yield_duration: Option<String>,
    pub deprecated: Option<bool>,
    pub deprecation_alternatives: Option<Vec<String>>,
    pub deprecation_reason: Option<String>,
    pub dynamic_properties: Vec<DynamicProperty>,
    pub dynamic_relationship: Option<DynamicRelationship>,
    pub explicit_restrictions: Option<Vec<Restriction>>,
    pub group: Option<String>,
    pub input_requirement: Option<ProcessorDefinitionInputRequirement>,
    pub multi_processor_use_cases: Vec<MultiProcessorUseCase>,
    pub primary_node_only: Option<bool>,
    pub property_descriptors: HashMap<String, PropertyDescriptor>,
    pub provided_api_implementations: Vec<DefinedType>,
    pub reads_attributes: Vec<Attribute>,
    pub restricted: Option<bool>,
    pub restricted_explanation: Option<String>,
    pub see_also: Option<Vec<String>>,
    pub side_effect_free: Option<bool>,
    pub stateful: Option<Stateful>,
    pub supported_relationships: Vec<Relationship>,
    pub supported_scheduling_strategies: Vec<String>,
    pub supports_batching: Option<bool>,
    pub supports_dynamic_properties: Option<bool>,
    pub supports_dynamic_relationships: Option<bool>,
    pub supports_sensitive_dynamic_properties: Option<bool>,
    pub system_resource_considerations: Vec<SystemResourceConsideration>,
    pub tags: Option<Vec<String>>,
    pub trigger_serially: Option<bool>,
    pub trigger_when_any_destination_available: Option<bool>,
    pub trigger_when_empty: Option<bool>,
    pub type_: Option<String>,
    pub type_description: Option<String>,
    pub use_cases: Vec<UseCase>,
    pub version: Option<String>,
    pub writes_attributes: Vec<Attribute>,

}

Processors provided in this bundle

JSON schema

{
 "description": "Processors provided in this bundle",
 "type": "object",
 "properties": {
   "additionalDetails": {
     "description": "Indicates if the component has additional details documentation",
     "type": "boolean"
   },
   "artifact": {
     "description": "The artifact name of the bundle that provides the referenced type.",
     "type": "string"
   },
   "buildInfo": {
     "$ref": "#/components/schemas/BuildInfo"
   },
   "defaultBulletinLevel": {
     "description": "The default bulletin level, such as WARN, INFO, DEBUG, etc.",
     "type": "string"
   },
   "defaultConcurrentTasksBySchedulingStrategy": {
     "description": "The default concurrent tasks for each scheduling strategy.",
     "type": "object",
     "additionalProperties": {
       "description": "The default concurrent tasks for each scheduling strategy.",
       "type": "integer",
       "format": "int32"
     }
   },
   "defaultPenaltyDuration": {
     "description": "The default penalty duration as a time period, such as \"30 sec\".",
     "type": "string"
   },
   "defaultSchedulingPeriodBySchedulingStrategy": {
     "description": "The default scheduling period for each scheduling strategy. The scheduling period is expected to be a time period, such as \"30 sec\".",
     "type": "object",
     "additionalProperties": {
       "description": "The default scheduling period for each scheduling strategy. The scheduling period is expected to be a time period, such as \"30 sec\".",
       "type": "string"
     }
   },
   "defaultSchedulingStrategy": {
     "description": "The default scheduling strategy for the processor.",
     "type": "string"
   },
   "defaultYieldDuration": {
     "description": "The default yield duration as a time period, such as \"1 sec\".",
     "type": "string"
   },
   "deprecated": {
     "description": "Whether or not the component has been deprecated",
     "type": "boolean"
   },
   "deprecationAlternatives": {
     "description": "If this component has been deprecated, this optional field provides alternatives to use",
     "type": "array",
     "items": {
       "description": "If this component has been deprecated, this optional field provides alternatives to use",
       "type": "string"
     },
     "uniqueItems": true
   },
   "deprecationReason": {
     "description": "If this component has been deprecated, this optional field can be used to provide an explanation",
     "type": "string"
   },
   "dynamicProperties": {
     "description": "Describes the dynamic properties supported by this component",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/DynamicProperty"
     }
   },
   "dynamicRelationship": {
     "$ref": "#/components/schemas/DynamicRelationship"
   },
   "explicitRestrictions": {
     "description": "Explicit restrictions that indicate a require permission to use the component",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/Restriction"
     },
     "uniqueItems": true
   },
   "group": {
     "description": "The group name of the bundle that provides the referenced type.",
     "type": "string"
   },
   "inputRequirement": {
     "description": "Any input requirements this processor has.",
     "type": "string",
     "enum": [
       "INPUT_REQUIRED",
       "INPUT_ALLOWED",
       "INPUT_FORBIDDEN"
     ]
   },
   "multiProcessorUseCases": {
     "description": "A list of use cases that have been documented that involve this Processor in conjunction with other Processors",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/MultiProcessorUseCase"
     }
   },
   "primaryNodeOnly": {
     "description": "Whether or not this processor should be scheduled only on the primary node in a cluster.",
     "type": "boolean"
   },
   "propertyDescriptors": {
     "description": "Descriptions of configuration properties applicable to this component.",
     "type": "object",
     "additionalProperties": {
       "$ref": "#/components/schemas/PropertyDescriptor"
     }
   },
   "providedApiImplementations": {
     "description": "If this type represents a provider for an interface, this lists the APIs it implements",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/DefinedType"
     }
   },
   "readsAttributes": {
     "description": "The FlowFile attributes this processor reads",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/Attribute"
     }
   },
   "restricted": {
     "description": "Whether or not the component has a general restriction",
     "type": "boolean"
   },
   "restrictedExplanation": {
     "description": "An optional description of the general restriction",
     "type": "string"
   },
   "seeAlso": {
     "description": "The names of other component types that may be related",
     "type": "array",
     "items": {
       "description": "The names of other component types that may be related",
       "type": "string"
     },
     "uniqueItems": true
   },
   "sideEffectFree": {
     "description": "Whether or not this processor is considered side-effect free. Side-effect free indicate that the processor's operations on FlowFiles can be safely repeated across process sessions.",
     "type": "boolean"
   },
   "stateful": {
     "$ref": "#/components/schemas/Stateful"
   },
   "supportedRelationships": {
     "description": "The supported relationships for this processor.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/Relationship"
     }
   },
   "supportedSchedulingStrategies": {
     "description": "The supported scheduling strategies, such as TIME_DRIVER, CRON, or EVENT_DRIVEN.",
     "type": "array",
     "items": {
       "description": "The supported scheduling strategies, such as TIME_DRIVER, CRON, or EVENT_DRIVEN.",
       "type": "string"
     }
   },
   "supportsBatching": {
     "description": "Whether or not this processor supports batching. If a Processor uses this annotation, it allows the Framework to batch calls to session commits, as well as allowing the Framework to return the same session multiple times.",
     "type": "boolean"
   },
   "supportsDynamicProperties": {
     "description": "Whether or not this component makes use of dynamic (user-set) properties.",
     "type": "boolean"
   },
   "supportsDynamicRelationships": {
     "description": "Whether or not this processor supports dynamic relationships.",
     "type": "boolean"
   },
   "supportsSensitiveDynamicProperties": {
     "description": "Whether or not this component makes use of sensitive dynamic (user-set) properties.",
     "type": "boolean"
   },
   "systemResourceConsiderations": {
     "description": "The system resource considerations for the given component",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/SystemResourceConsideration"
     }
   },
   "tags": {
     "description": "The tags associated with this type",
     "type": "array",
     "items": {
       "description": "The tags associated with this type",
       "type": "string"
     },
     "uniqueItems": true
   },
   "triggerSerially": {
     "description": "Whether or not this processor should be triggered serially (i.e. no concurrent execution).",
     "type": "boolean"
   },
   "triggerWhenAnyDestinationAvailable": {
     "description": "Whether or not this processor should be triggered when any destination queue has room.",
     "type": "boolean"
   },
   "triggerWhenEmpty": {
     "description": "Whether or not this processor should be triggered when incoming queues are empty.",
     "type": "boolean"
   },
   "type": {
     "description": "The fully-qualified class type",
     "type": "string"
   },
   "typeDescription": {
     "description": "The description of the type.",
     "type": "string"
   },
   "useCases": {
     "description": "A list of use cases that have been documented for this Processor",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/UseCase"
     }
   },
   "version": {
     "description": "The version of the bundle that provides the referenced type.",
     "type": "string"
   },
   "writesAttributes": {
     "description": "The FlowFile attributes this processor writes/updates",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/Attribute"
     }
   }
 }
}

Fields
additional_details: Option<bool>

Indicates if the component has additional details documentation

artifact: Option<String>

The artifact name of the bundle that provides the referenced type.

build_info: Option<BuildInfo>default_bulletin_level: Option<String>

The default bulletin level, such as WARN, INFO, DEBUG, etc.

default_concurrent_tasks_by_scheduling_strategy: HashMap<String, i32>

The default concurrent tasks for each scheduling strategy.

default_penalty_duration: Option<String>

The default penalty duration as a time period, such as “30 sec”.

default_scheduling_period_by_scheduling_strategy: HashMap<String, String>

The default scheduling period for each scheduling strategy. The scheduling period is expected to be a time period, such as “30 sec”.

default_scheduling_strategy: Option<String>

The default scheduling strategy for the processor.

default_yield_duration: Option<String>

The default yield duration as a time period, such as “1 sec”.

deprecated: Option<bool>

Whether or not the component has been deprecated

deprecation_alternatives: Option<Vec<String>>

If this component has been deprecated, this optional field provides alternatives to use

deprecation_reason: Option<String>

If this component has been deprecated, this optional field can be used to provide an explanation

dynamic_properties: Vec<DynamicProperty>

Describes the dynamic properties supported by this component

dynamic_relationship: Option<DynamicRelationship>explicit_restrictions: Option<Vec<Restriction>>

Explicit restrictions that indicate a require permission to use the component

group: Option<String>

The group name of the bundle that provides the referenced type.

input_requirement: Option<ProcessorDefinitionInputRequirement>

Any input requirements this processor has.

multi_processor_use_cases: Vec<MultiProcessorUseCase>

A list of use cases that have been documented that involve this Processor in conjunction with other Processors

primary_node_only: Option<bool>

Whether or not this processor should be scheduled only on the primary node in a cluster.

property_descriptors: HashMap<String, PropertyDescriptor>

Descriptions of configuration properties applicable to this component.

provided_api_implementations: Vec<DefinedType>

If this type represents a provider for an interface, this lists the APIs it implements

reads_attributes: Vec<Attribute>

The FlowFile attributes this processor reads

restricted: Option<bool>

Whether or not the component has a general restriction

restricted_explanation: Option<String>

An optional description of the general restriction

see_also: Option<Vec<String>>

The names of other component types that may be related

side_effect_free: Option<bool>

Whether or not this processor is considered side-effect free. Side-effect free indicate that the processor’s operations on FlowFiles can be safely repeated across process sessions.

stateful: Option<Stateful>supported_relationships: Vec<Relationship>

The supported relationships for this processor.

supported_scheduling_strategies: Vec<String>

The supported scheduling strategies, such as TIME_DRIVER, CRON, or EVENT_DRIVEN.

supports_batching: Option<bool>

Whether or not this processor supports batching. If a Processor uses this annotation, it allows the Framework to batch calls to session commits, as well as allowing the Framework to return the same session multiple times.

supports_dynamic_properties: Option<bool>

Whether or not this component makes use of dynamic (user-set) properties.

supports_dynamic_relationships: Option<bool>

Whether or not this processor supports dynamic relationships.

supports_sensitive_dynamic_properties: Option<bool>

Whether or not this component makes use of sensitive dynamic (user-set) properties.

system_resource_considerations: Vec<SystemResourceConsideration>

The system resource considerations for the given component

tags: Option<Vec<String>>

The tags associated with this type

trigger_serially: Option<bool>

Whether or not this processor should be triggered serially (i.e. no concurrent execution).

trigger_when_any_destination_available: Option<bool>

Whether or not this processor should be triggered when any destination queue has room.

trigger_when_empty: Option<bool>

Whether or not this processor should be triggered when incoming queues are empty.

type_: Option<String>

The fully-qualified class type

type_description: Option<String>

The description of the type.

use_cases: Vec<UseCase>

A list of use cases that have been documented for this Processor

version: Option<String>

The version of the bundle that provides the referenced type.

writes_attributes: Vec<Attribute>

The FlowFile attributes this processor writes/updates

Implementations

impl ProcessorDefinition

pub fn builder() -> ProcessorDefinition