nifi_rs::proxy::v260::api

Struct VersionedConnection

pub struct VersionedConnection {

    pub back_pressure_data_size_threshold: Option<String>,
    pub back_pressure_object_threshold: Option<i64>,
    pub bends: Vec<Position>,
    pub comments: Option<String>,
    pub component_type: Option<VersionedConnectionComponentType>,
    pub destination: Option<ConnectableComponent>,
    pub flow_file_expiration: Option<String>,
    pub group_identifier: Option<String>,
    pub identifier: Option<String>,
    pub instance_identifier: Option<String>,
    pub label_index: Option<i32>,
    pub load_balance_compression: Option<String>,
    pub load_balance_strategy: Option<String>,
    pub name: Option<String>,
    pub partitioning_attribute: Option<String>,
    pub position: Option<Position>,
    pub prioritizers: Vec<String>,
    pub selected_relationships: Option<Vec<String>>,
    pub source: Option<ConnectableComponent>,
    pub z_index: Option<i64>,

}

The Connections

JSON schema

{
 "description": "The Connections",
 "type": "object",
 "properties": {
   "backPressureDataSizeThreshold": {
     "description": "The object data size threshold for determining when back pressure is applied. Updating this value is a passive change in the sense that it won't impact whether existing files over the limit are affected but it does help feeder processors to stop pushing too much into this work queue.",
     "type": "string"
   },
   "backPressureObjectThreshold": {
     "description": "The object count threshold for determining when back pressure is applied. Updating this value is a passive change in the sense that it won't impact whether existing files over the limit are affected but it does help feeder processors to stop pushing too much into this work queue.",
     "type": "integer",
     "format": "int64"
   },
   "bends": {
     "description": "The bend points on the connection.",
     "type": "array",
     "items": {
       "$ref": "#/components/schemas/Position"
     }
   },
   "comments": {
     "description": "The user-supplied comments for the component",
     "type": "string"
   },
   "componentType": {
     "type": "string",
     "enum": [
       "CONNECTION",
       "PROCESSOR",
       "PROCESS_GROUP",
       "REMOTE_PROCESS_GROUP",
       "INPUT_PORT",
       "OUTPUT_PORT",
       "REMOTE_INPUT_PORT",
       "REMOTE_OUTPUT_PORT",
       "FUNNEL",
       "LABEL",
       "CONTROLLER_SERVICE",
       "REPORTING_TASK",
       "FLOW_ANALYSIS_RULE",
       "PARAMETER_CONTEXT",
       "PARAMETER_PROVIDER",
       "FLOW_REGISTRY_CLIENT"
     ]
   },
   "destination": {
     "$ref": "#/components/schemas/ConnectableComponent"
   },
   "flowFileExpiration": {
     "description": "The amount of time a flow file may be in the flow before it will be automatically aged out of the flow. Once a flow file reaches this age it will be terminated from the flow the next time a processor attempts to start work on it.",
     "type": "string"
   },
   "groupIdentifier": {
     "description": "The ID of the Process Group that this component belongs to",
     "type": "string"
   },
   "identifier": {
     "description": "The component's unique identifier",
     "type": "string"
   },
   "instanceIdentifier": {
     "description": "The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component",
     "type": "string"
   },
   "labelIndex": {
     "description": "The index of the bend point where to place the connection label.",
     "type": "integer",
     "format": "int32"
   },
   "loadBalanceCompression": {
     "description": "Whether or not compression should be used when transferring FlowFiles between nodes Possible returned values: DO_NOT_COMPRESS, COMPRESS_ATTRIBUTES_ONLY, COMPRESS_ATTRIBUTES_AND_CONTENT. See LoadBalanceCompression.class for more details.",
     "type": "string"
   },
   "loadBalanceStrategy": {
     "description": "The Strategy to use for load balancing data across the cluster, or null, if no Load Balance Strategy has been specified. Possible returned values: DO_NOT_LOAD_BALANCE, PARTITION_BY_ATTRIBUTE, ROUND_ROBIN, SINGLE_NODE. See LoadBalanceStrategy.class for more details.",
     "type": "string"
   },
   "name": {
     "description": "The component's name",
     "type": "string"
   },
   "partitioningAttribute": {
     "description": "The attribute to use for partitioning data as it is load balanced across the cluster. If the Load Balance Strategy is configured to use PARTITION_BY_ATTRIBUTE, the value returned by this method is the name of the FlowFile Attribute that will be used to determine which node in the cluster should receive a given FlowFile. If the Load Balance Strategy is unset or is set to any other value, the Partitioning Attribute has no effect.",
     "type": "string"
   },
   "position": {
     "$ref": "#/components/schemas/Position"
   },
   "prioritizers": {
     "description": "The comparators used to prioritize the queue.",
     "type": "array",
     "items": {
       "description": "The comparators used to prioritize the queue.",
       "type": "string"
     }
   },
   "selectedRelationships": {
     "description": "The selected relationship that comprise the connection.",
     "type": "array",
     "items": {
       "description": "The selected relationship that comprise the connection.",
       "type": "string"
     },
     "uniqueItems": true
   },
   "source": {
     "$ref": "#/components/schemas/ConnectableComponent"
   },
   "zIndex": {
     "description": "The z index of the connection.",
     "type": "integer",
     "format": "int64"
   }
 }
}

Fields
back_pressure_data_size_threshold: Option<String>

The object data size threshold for determining when back pressure is applied. Updating this value is a passive change in the sense that it won’t impact whether existing files over the limit are affected but it does help feeder processors to stop pushing too much into this work queue.

back_pressure_object_threshold: Option<i64>

The object count threshold for determining when back pressure is applied. Updating this value is a passive change in the sense that it won’t impact whether existing files over the limit are affected but it does help feeder processors to stop pushing too much into this work queue.

bends: Vec<Position>

The bend points on the connection.

comments: Option<String>

The user-supplied comments for the component

component_type: Option<VersionedConnectionComponentType>destination: Option<ConnectableComponent>flow_file_expiration: Option<String>

The amount of time a flow file may be in the flow before it will be automatically aged out of the flow. Once a flow file reaches this age it will be terminated from the flow the next time a processor attempts to start work on it.

group_identifier: Option<String>

The ID of the Process Group that this component belongs to

identifier: Option<String>

The component’s unique identifier

instance_identifier: Option<String>

The instance ID of an existing component that is described by this VersionedComponent, or null if this is not mapped to an instantiated component

label_index: Option<i32>

The index of the bend point where to place the connection label.

load_balance_compression: Option<String>

Whether or not compression should be used when transferring FlowFiles between nodes Possible returned values: DO_NOT_COMPRESS, COMPRESS_ATTRIBUTES_ONLY, COMPRESS_ATTRIBUTES_AND_CONTENT. See LoadBalanceCompression.class for more details.

load_balance_strategy: Option<String>

The Strategy to use for load balancing data across the cluster, or null, if no Load Balance Strategy has been specified. Possible returned values: DO_NOT_LOAD_BALANCE, PARTITION_BY_ATTRIBUTE, ROUND_ROBIN, SINGLE_NODE. See LoadBalanceStrategy.class for more details.

name: Option<String>

The component’s name

partitioning_attribute: Option<String>

The attribute to use for partitioning data as it is load balanced across the cluster. If the Load Balance Strategy is configured to use PARTITION_BY_ATTRIBUTE, the value returned by this method is the name of the FlowFile Attribute that will be used to determine which node in the cluster should receive a given FlowFile. If the Load Balance Strategy is unset or is set to any other value, the Partitioning Attribute has no effect.

position: Option<Position>prioritizers: Vec<String>

The comparators used to prioritize the queue.

selected_relationships: Option<Vec<String>>

The selected relationship that comprise the connection.

source: Option<ConnectableComponent>z_index: Option<i64>

The z index of the connection.

Implementations

impl VersionedConnection

pub fn builder() -> VersionedConnection