nifi_rs::proxy::v260

Module parameter_context

Parameter Context Module

Provides high-level bindings for the NiFi “parameter-contexts” API endpoint.

This module allows for creating, reading, and updating Parameter Contexts, which are collections of parameters that can be shared across Process Groups.

Structs

ParameterContext

A service for interacting with NiFi’s Parameter Context endpoints.